refactor(rewards): convert EvaaMasterRewardsConfig to based EvaaRewardsConfig
