+++ src/constants/pools.ts
@@ -1,6 +1,6 @@
 import { Address } from '@ton/core';
 import { PoolConfig } from '../types/Master';
-import { EvaaMasterRewardsConfig } from '../types/MasterRewards';
+import { EvaaRewardsConfig } from '../types/MasterRewards';
 import {
     CATI_MAINNET,
     DOGS_MAINNET,
@@ -86,7 +86,7 @@ export const MAINNET_ALTS_POOL_CONFIG: PoolConfig = {
     lendingCode: LENDING_CODE,
 };
 
-export const TESTNET_MASTER_REWARD_CONFIG: EvaaMasterRewardsConfig = {
+export const TESTNET_MASTER_REWARD_CONFIG: EvaaRewardsConfig = {
     adminAddress: EVAA_REWARDS_MASTER_TESTNET,
     evaaMasterAddress: new Address(0, Buffer.alloc(32, 0)),
     rewardMasterCode: EVAA_REWARDS_MASTER_CODE_TESTNET,
@@ -96,7 +96,7 @@ export const TESTNET_MASTER_REWARD_CONFIG: EvaaMasterRewardsConfig = {
     publicKey: Buffer.from('c38be45e033419b49cd22c97fbd23610fbc851a931208f8b4a1817dd11cb63ac', 'hex'), // adminAddress publicKey
 };
 
-export const TESTNET_MASTER_EUSDT_REWARD_CONFIG: EvaaMasterRewardsConfig = {
+export const TESTNET_MASTER_EUSDT_REWARD_CONFIG: EvaaRewardsConfig = {
     adminAddress: EVAA_REWARDS_MASTER_TESTNET,
     evaaMasterAddress: new Address(0, Buffer.alloc(32, 0)),
     rewardMasterCode: EVAA_REWARDS_MASTER_CODE_TESTNET,
