+++ src/rewards/EvaaRewards.ts
@@ -1,21 +1,21 @@
 import { Address, Contract, ContractProvider } from '@ton/ton';
-import { EvaaUserRewardsConfig } from '../types/UserRewards';
+import { EvaaRewardsConfig } from '../types/MasterRewards';
 import { RewardUser } from './RewardUser';
 
 export class EvaaUserRewards implements Contract {
     readonly address: Address;
     constructor(
         readonly userAddress: Address,
-        readonly config: EvaaUserRewardsConfig,
+        readonly config: Omit<EvaaRewardsConfig, 'userAddress'>,
     ) {
         this.address = this.userAddress;
     }
 
-    openUserContract(provider: ContractProvider) {
+    openContract(provider: ContractProvider) {
         return provider.open(
             RewardUser.createFromConfig({
                 asset: this.config.asset,
-                rewardMasterAddress: this.config.rewardMasterAddress,
+                rewardMasterAddress: this.config.adminAddress,
                 rewardUserCode: this.config.rewardUserCode,
                 userAddress: this.userAddress,
                 publicKey: this.config.publicKey,
